Move stubs to their own namespace
